style: A lot of sonarlint related code changes

+++ GeneaGrab.Core/Helpers/Zoomify.cs
@@ -12,7 +12,7 @@ namespace GeneaGrab.Core.Helpers
         /// <summary>Returns the maximum zoom level</summary>
         // Each zoom level multiplies the size of the image by two. The zoom level 0 corresponds to the entire image contained in a single tile.
         // ReSharper disable once PossibleLossOfFraction
-        public static int CalculateIndex(RPage page) => Math.Max((int)Math.Ceiling(Math.Log(Math.Max(page.Width, page.Height) / page.TileSize.GetValueOrDefault(256)) / Math.Log(2)), 0);
+        public static int CalculateIndex(RPage page) => Math.Max((int)Math.Ceiling(Math.Log(Math.Max(page.Width, page.Height) / (double)page.TileSize.GetValueOrDefault(256)) / Math.Log(2)), 0);
 
         /// <summary>Returns the properties of the image</summary>
         /// <param name="baseURL">Image root url</param>
